requirements.txt
django>=4.0
djangorestframework
django-environ
drf-yasg
celery
redis
gunicorn
psycopg2-binary
djangorestframework-simplejwt
sentry-sdk
python-decouple
Pillow>=9.0
django-cors-headers